requirements.txt
alabaster==0.7.13
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.12.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
django-allauth==0.52.0
django-bootstrap-v5==1.0.11
django-environ==0.9.0
django-sendgrid-v5==1.2.2
docutils==0.20.1
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.11.2
filelock==3.9.0
hyperlink==21.0.0
identify==2.5.20
idna==3.4
imagesize==1.4.1
importlib-metadata==6.7.0
incremental==22.10.0
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.4
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.2.0
pre-commit==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytest==7.3.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.7
python-http-client==3.3.7
python3-openid==3.2.0
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
sendgrid==6.10.0
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.20.0
zipp==3.15.0
zope.interface==5.5.2
gunicorn==21.2.0
